Setup vim9 + copilot
Some minor other changes, but primarily this sets up a build from
source of the latest vim version, needed to get vim version 9 as the
current vim.
With that done, it is possible to setup GitHub Copilot. Since this
will likely conflict heavily with YCM, it is enabled conditionally as
a YCM replacement, based on the vim version number.

In support of copilot, some steps to install fnm are also added.